backend/pyproject.toml
[project]
name = "attendance-tracking-sys-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
